fix: move BorderStack component to the layout folder
