+++ src/design/layout/borderStack/styles.ts
