+++ src/pages/missions/tabs/requestsInProgressTab/tools/details/requestsInProcess/index.tsx
@@ -5,7 +5,7 @@ import { ModalWrapper } from "@design/modals/modalWrapper";
 import { DetailBox } from "@design/feedback/detailBox";
 import { IEntry } from "@ptypes/design/table/IEntry";
 import { basic } from "@design/tokens";
-import { BorderStack } from "@design/modals/borderStack";
+import { BorderStack } from "@design/layout/borderStack";
 import { detailsRequestInProgressModal } from "@config/requestsInProgressTab/details/detailsRequestInProgressModal";
 import { IRequestsInProcess } from "@ptypes/requestsInProgress/IRequestsInProcess";
 import { EComponentAppearance } from "@enum/appearances";
