+++ src/pages/missions/tools/detailsModal/detailsMissionModal/interface.tsx
@@ -1,5 +1,5 @@
 import { ModalWrapper } from "@design/modals/modalWrapper";
-import { BorderStack } from "@design/modals/borderStack";
+import { BorderStack } from "@design/layout/borderStack";
 import { basic } from "@design/tokens";
 import RequestTitleSection from "@pages/positions/tabs/requestsInProgressTab/tools/details/requestsInProcess/requestTitleSection";
 import { IDetailsMissionModalUI } from "@ptypes/missions/IDetailsMissionModalUI";
